# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Healthy App",
        "description": "API documentation for Healthy",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/meals": {
            "get": {
                "tags": [
                    "Meals"
                ],
                "summary": "Get a list of meals",
                "operationId": "4d2d017ae2b8ac8cecf16eb53c6d5fd9",
                "parameters": [
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Filter by availability",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "dietType",
                        "in": "query",
                        "description": "Filter by diet type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ownerId",
                        "in": "query",
                        "description": "Filter by owner ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MealResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Meals"
                ],
                "summary": "Create a new meal",
                "operationId": "06b46f308f52e05ab22bd7c12ce2d15a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMealRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Meal created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MealResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/meals/{id}": {
            "get": {
                "tags": [
                    "Meals"
                ],
                "summary": "Get a specific meal",
                "operationId": "7ddd9781693c95914584c78006bb752a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the meal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MealResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Meals"
                ],
                "summary": "Update a specific meal",
                "operationId": "6fe35e91135dd0abcc8d3f701be49aa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the meal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMealRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Meal updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MealResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Meals"
                ],
                "summary": "Delete a specific meal",
                "operationId": "ad04fdbda4295360fbd349f436e8a974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the meal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Meal deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Meal deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreMealRequest": {
                "title": "Store Meal Request",
                "description": "Request body for storing a new meal",
                "required": [
                    "ownerId",
                    "title",
                    "price"
                ],
                "properties": {
                    "ownerId": {
                        "description": "ID of the owner",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title of the meal",
                        "type": "string",
                        "maxLength": 150,
                        "example": "Delicious Pasta"
                    },
                    "description": {
                        "description": "Description of the meal",
                        "type": "string",
                        "example": "A classic Italian pasta dish"
                    },
                    "price": {
                        "description": "Price of the meal",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 12.99
                    },
                    "isAvailable": {
                        "description": "Indicates if the meal is available",
                        "type": "boolean",
                        "example": true
                    },
                    "availableFrom": {
                        "description": "Time from which the meal is available",
                        "type": "string",
                        "format": "time",
                        "example": "12:00:00"
                    },
                    "availableTo": {
                        "description": "Time until which the meal is available",
                        "type": "string",
                        "format": "time",
                        "example": "18:00:00"
                    },
                    "dietType": {
                        "description": "Type of diet the meal is suitable for",
                        "type": "string",
                        "enum": [
                            "keto",
                            "low_carb",
                            "vegetarian",
                            "vegan",
                            "paleo",
                            "balanced"
                        ],
                        "example": "vegetarian"
                    },
                    "allergenIds": {
                        "description": "List of allergen IDs",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    },
                    "ingredients": {
                        "description": "List of ingredients with their quantities and units",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "quantity": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 100
                                },
                                "unit": {
                                    "type": "string",
                                    "enum": [
                                        "tbsp",
                                        "g",
                                        "piece",
                                        "l",
                                        "ml"
                                    ],
                                    "example": "g"
                                }
                            },
                            "type": "object"
                        },
                        "example": [
                            {
                                "id": 1,
                                "quantity": 100,
                                "unit": "g"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "UpdateMealRequest": {
                "properties": {
                    "ownerId": {
                        "description": "The ID of the owner",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "The title of the meal",
                        "type": "string",
                        "maxLength": 150,
                        "example": "Delicious Meal"
                    },
                    "description": {
                        "description": "The description of the meal",
                        "type": "string",
                        "example": "A very tasty meal"
                    },
                    "price": {
                        "description": "The price of the meal",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 9.99
                    },
                    "isAvailable": {
                        "description": "Whether the meal is available",
                        "type": "boolean",
                        "example": true
                    },
                    "availableFrom": {
                        "description": "The time from which the meal is available",
                        "type": "string",
                        "format": "time",
                        "example": "08:00:00"
                    },
                    "availableTo": {
                        "description": "The time until which the meal is available",
                        "type": "string",
                        "format": "time",
                        "example": "18:00:00"
                    },
                    "dietType": {
                        "description": "The type of diet",
                        "type": "string",
                        "enum": [
                            "keto",
                            "low_carb",
                            "vegetarian",
                            "vegan",
                            "paleo",
                            "balanced"
                        ],
                        "example": "vegetarian"
                    },
                    "allergenIds": {
                        "description": "Array of allergen IDs",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    },
                    "ingredients": {
                        "description": "Array of ingredients with their quantities and units",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "quantity": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 2.5
                                },
                                "unit": {
                                    "type": "string",
                                    "enum": [
                                        "tbsp",
                                        "g",
                                        "piece",
                                        "l"
                                    ],
                                    "example": "g"
                                }
                            },
                            "type": "object"
                        },
                        "example": [
                            {
                                "id": 1,
                                "quantity": 2.5,
                                "unit": "g"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "AllergenResource": {
                "title": "Allergen Resource",
                "description": "Allergen resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the allergen",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the allergen",
                        "type": "string"
                    },
                    "mealsCount": {
                        "description": "The count of meals associated with the allergen",
                        "type": "integer"
                    },
                    "createdAt": {
                        "description": "The date when the allergen was created",
                        "type": "string",
                        "format": "date"
                    },
                    "updatedAt": {
                        "description": "The date when the allergen was last updated",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
            "MealResource": {
                "title": "Meal Resource",
                "description": "Meal resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the meal",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the meal",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the meal",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the meal in currency",
                        "type": "number",
                        "format": "float"
                    },
                    "isAvailable": {
                        "description": "Indicates if the meal is available",
                        "type": "boolean"
                    },
                    "availability": {
                        "description": "The availability time frame of the meal",
                        "properties": {
                            "from": {
                                "description": "The start time of availability",
                                "type": "string",
                                "format": "date-time"
                            },
                            "to": {
                                "description": "The end time of availability",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    },
                    "dietType": {
                        "description": "The diet type of the meal",
                        "type": "string"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "allergens": {
                        "description": "List of allergens associated with the meal",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AllergenResource"
                        }
                    },
                    "ingredients": {
                        "description": "List of ingredients in the meal",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IngredientResource"
                        }
                    },
                    "createdAt": {
                        "description": "The creation date of the meal",
                        "type": "string",
                        "format": "date"
                    },
                    "updatedAt": {
                        "description": "The last update date of the meal",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "Allergen Resource",
                "description": "Allergen resource representation",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}